Refactor: Separated the  pages into individual files for better organization
